Renaming from PusherDotNet to PusherRESTDotNet to indicate it is a REST library.
Adding nuget package.
