--- PusherDotNet.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PusherDotNet.Tests", "PusherDotNet.Tests\PusherDotNet.Tests.csproj", "{29ADB209-999D-4A86-A006-01609B6B670D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PusherDotNet", "PusherDotNet\PusherDotNet.csproj", "{A4889613-5DF5-41B9-8B92-D9647545872D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{29ADB209-999D-4A86-A006-01609B6B670D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29ADB209-999D-4A86-A006-01609B6B670D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29ADB209-999D-4A86-A006-01609B6B670D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29ADB209-999D-4A86-A006-01609B6B670D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A4889613-5DF5-41B9-8B92-D9647545872D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A4889613-5DF5-41B9-8B92-D9647545872D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A4889613-5DF5-41B9-8B92-D9647545872D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A4889613-5DF5-41B9-8B92-D9647545872D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
