+++ PusherRESTDotNet.Tests/Properties/AssemblyInfo.cs
