+++ PusherRESTDotNet.Tests/PusherProviderTests.cs
@@ -3,7 +3,7 @@ using System.Configuration;
 using System.Net;
 using NUnit.Framework;
 
-namespace PusherDotNet.Tests
+namespace PusherRESTDotNet.Tests
 {
 	[TestFixture]
 	public class PusherProviderTests
