+++ PusherRESTDotNet.sln
@@ -1,0 +1,32 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PusherRESTDotNet", "PusherRESTDotNet\PusherRESTDotNet.csproj", "{FEDAD44D-208E-4F99-BE08-6C1257B59D4A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PusherRESTDotNet.Tests", "PusherRESTDotNet.Tests\PusherRESTDotNet.Tests.csproj", "{29ADB209-999D-4A86-A006-01609B6B670D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Files", "Files", "{13FCFB53-1898-4AE9-B229-F917B8F52DB2}"
+	ProjectSection(SolutionItems) = preProject
+		PusherRESTDotNet\PusherRESTDotNet.nuspec = PusherRESTDotNet\PusherRESTDotNet.nuspec
+		README.txt = README.txt
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FEDAD44D-208E-4F99-BE08-6C1257B59D4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FEDAD44D-208E-4F99-BE08-6C1257B59D4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FEDAD44D-208E-4F99-BE08-6C1257B59D4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FEDAD44D-208E-4F99-BE08-6C1257B59D4A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29ADB209-999D-4A86-A006-01609B6B670D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29ADB209-999D-4A86-A006-01609B6B670D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29ADB209-999D-4A86-A006-01609B6B670D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29ADB209-999D-4A86-A006-01609B6B670D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
