+++ PusherRESTDotNet/Properties/AssemblyInfo.cs
