+++ PusherRESTDotNet/PusherRESTDotNet.nuspec
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
+  <metadata>
+    <id>$id$</id>
+    <version>$version$</version>
+    <authors>Graham Scott, Phil Leggetter</authors>
+    <owners>$author$</owners>
+    <projectUrl>https://github.com/leggetter/pusher-rest-dotnet</projectUrl>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>The Pusher (http://pusher.com) REST library provides a simple way of pushing realtime messages into Pusher via the Pusher REST API.</description>
+    <tags>realtime push websockets REST API Pusher</tags>
+  </metadata>
+</package>
